requirements_txt.txt
line-bot-sdk==2.4.2
Flask==2.3.3
schedule==1.2.0
gunicorn==21.2.0